Refactor api::Events processing
